增加入参校验

+++ translationtoolservice/src/main/java/com/shr/translationtoolservice/controller/ConfigManageController.java
@@ -9,10 +9,16 @@ import com.shr.translationtoolservice.service.impl.ConfigManageServiceImpl;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import lombok.extern.slf4j.Slf4j;
+import org.junit.platform.commons.util.StringUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.util.CollectionUtils;
 import org.springframework.web.bind.annotation.*;
+import org.yaml.snakeyaml.util.ArrayUtils;
+import sun.security.util.ArrayUtil;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
 
 /**
  * @ClassName ConfigManageController
@@ -48,6 +54,9 @@ public class ConfigManageController extends BaseController {
     @ApiOperation("新增用户")
     @CrossOrigin
     public HttpResponse<String> addUser(@RequestBody ConfigResUser user) {
+        if (Objects.isNull(user) || StringUtils.isBlank(user.getUserName())) {
+            return checkResult(ErrorCodeList.INPUT_IS_NULL);
+        }
         String userRes = configManageService.addUser(user);
 
         return checkResult(userRes);
@@ -56,17 +65,21 @@ public class ConfigManageController extends BaseController {
     @PostMapping("/updateUserInfo")
     @ApiOperation("编辑用户信息")
     @CrossOrigin
-    public HttpResponse<Integer> updateUserInfo(@RequestBody ConfigResUser user) {
-        Integer res = configManageService.updateUserInfo(user);
-        return checkResult(res);
+    public HttpResponse<String> updateUserInfo(@RequestBody ConfigResUser user) {
+        if (Objects.isNull(user)) {
+            return checkResult(ErrorCodeList.INPUT_IS_NULL);
+        }
+        return checkResult(configManageService.updateUserInfo(user));
     }
 
     @PostMapping("/deleteUser")
     @ApiOperation("删除用户信息")
     @CrossOrigin
-    public HttpResponse<Integer> deleteUserInfoByList(@RequestBody List<String> idList) {
-        Integer res = configManageService.deleteUserInfoByList(idList);
-        return checkResult(res);
+    public HttpResponse<String> deleteUserInfoByList(@RequestBody List<String> idList) {
+        if (CollectionUtils.isEmpty(idList)) {
+            return checkResult(ErrorCodeList.INPUT_IS_NULL);
+        }
+        return checkResult(configManageService.deleteUserInfoByList(idList));
     }
 
 
@@ -90,59 +103,110 @@ public class ConfigManageController extends BaseController {
     @PostMapping("/deleteRoleInfo")
     @ApiOperation("删除角色信息")
     @CrossOrigin
-    public HttpResponse<Integer> deleteRoleInfo(@RequestBody List<String> roleIDs) {
-
+    public HttpResponse<String> deleteRoleInfo(@RequestBody List<String> roleIDs) {
+        if (CollectionUtils.isEmpty(roleIDs)) {
+            return checkResult(ErrorCodeList.INPUT_IS_NULL);
+        }
         return checkResult(configManageService.deleteRoleInfo(roleIDs));
     }
 
     @PostMapping("/updateRoleInfo")
     @ApiOperation("编辑角色信息")
     @CrossOrigin
-    public HttpResponse<Integer> updateRoleInfo(Role role) {
-
+    public HttpResponse<String> updateRoleInfo(Role role) {
+        if (Objects.isNull(role)) {
+            return checkResult(ErrorCodeList.INPUT_IS_NULL);
+        }
         return checkResult(configManageService.updateRoleInfo(role));
     }
 
     @PostMapping("/addRoleInfo")
     @ApiOperation("新增角色信息")
     @CrossOrigin
-    public HttpResponse<Integer> addRoleInfo(Role role) {
-
-        return checkResult(configManageService.insertSelective(role));
+    public HttpResponse<String> addRoleInfo(Role role) {
+        if (Objects.isNull(role) || StringUtils.isBlank(role.getRoleName())) {
+            return checkResult(ErrorCodeList.INPUT_IS_NULL);
+        }
+        return checkResult(configManageService.insertRoleInfo(role));
     }
 
     @PostMapping("/bindRoleInfo")
     @ApiOperation("绑定角色信息")
     @CrossOrigin
-    public HttpResponse<Integer> bindRoleInfo(ConfigResUser configResUser) {
-
+    public HttpResponse<String> bindRoleInfo(ConfigResUser configResUser) {
+        if (Objects.isNull(configResUser) || StringUtils.isNotBlank(configResUser.getId())) {
+            return checkResult(ErrorCodeList.INPUT_IS_NULL);
+        }
         return checkResult(configManageService.bindRoleInfo(configResUser));
     }
 
     @PostMapping("/bindPermission")
     @ApiOperation("绑定权限信息")
     @CrossOrigin
-    public HttpResponse<Integer> bindPermission(@RequestBody RoleAuthority roleAuthority) {
+    public HttpResponse<String> bindPermission(@RequestBody RoleAuthority roleAuthority) {
 
+        if (Objects.isNull(roleAuthority)){
+            return checkResult(ErrorCodeList.INPUT_IS_NULL);
+        }
         return checkResult(configManageService.bindPermission(roleAuthority));
     }
-}
 
-/*
     @PostMapping("/queryVersionInfo")
     @ApiOperation("查询版本信息")
     @PassToken
     @CrossOrigin
-    public HttpResponse<ResponseListModel> queryVersionInfo( String versionName,
-                                                             @RequestParam(value = "pageIndex", defaultValue = "1") Integer pageIndex,
-                                                             @RequestParam(value = "pageSize", defaultValue = "20") Integer pageSize){
+    public HttpResponse<ResponseListModel> queryVersionInfo(String versionName,
+                                                            @RequestParam(value = "pageIndex", defaultValue = "1") Integer pageIndex,
+                                                            @RequestParam(value = "pageSize", defaultValue = "20") Integer pageSize) {
         ResponseListModel<EntryVersion> result = new ResponseListModel<>();
-        List<EntryVersion> roleRes = configManageService.queryVersionInfo(versionName,pageIndex,pageSize);
-        i//nt total = configManageService.getVersionTotaNum(versionName);
+        List<EntryVersion> roleRes = configManageService.queryVersionInfo(versionName, pageIndex, pageSize);
+        int total = configManageService.getVersionTotaNum(versionName);
         result.setList(roleRes);
 
         result.setTotalNum(total);
 
         return checkResult(result);
-    }}
+    }
+
+    @PostMapping("/updateVersionInfo")
+    @ApiOperation("编辑版本信息")
+    @PassToken
+    @CrossOrigin
+    //TODO
+    public HttpResponse<String> updateVersionInfo(EntryVersion entryVersion) {
+        if (Objects.isNull(entryVersion) || StringUtils.isBlank(entryVersion.getId())){
+            return checkResult(ErrorCodeList.INPUT_IS_NULL);
+        }
+        return checkResult(configManageService.updateVersionInfo(entryVersion));
+    }
+
+    @PostMapping("/deleteVersionInfo")
+    @ApiOperation("删除版本信息")
+    @PassToken
+    @CrossOrigin
+    //TODO
+    public HttpResponse<String> deleteVersionInfo(@RequestBody List<String> idList) {
+        if (CollectionUtils.isEmpty(idList)){
+            return checkResult(ErrorCodeList.INPUT_IS_NULL);
+        }
+        return checkResult(configManageService.deleteVersionInfo(idList));
+    }
+
+    @PostMapping("/addVersionInfo")
+    @ApiOperation("新增版本信息")
+    @PassToken
+    @CrossOrigin
+    //TODO
+    public HttpResponse<String> addVersionInfo(EntryVersion entryVersion) {
+        if (Objects.isNull(entryVersion) || StringUtils.isBlank(entryVersion.getName())){
+            return checkResult(ErrorCodeList.INPUT_IS_NULL);
+        }
+        return checkResult(configManageService.addVersionInfo(entryVersion));
+    }
+
+
+}
+
+/*
+}
 */
